Add files for Initializable and Representable implementations
