+++ Sources/StructuredData.swift
@@ -32,6 +32,8 @@ public enum StructuredDataError: ErrorProtocol {
     case cannotInitialize(type: Any.Type, from: Any.Type)
 }
 
+protocol StructuredDataConvertible : StructuredDataInitializable, StructuredDataFallibleRepresentable {}
+
 // MARK: Parser/Serializer Protocols
 public protocol StructuredDataParser {
     func parse(_ data: Data) throws -> StructuredData
@@ -45,127 +47,6 @@ extension StructuredDataParser {
 
 public protocol StructuredDataSerializer {
     func serialize(_ structuredData: StructuredData) throws -> Data
-}
-
-// MARK: StructuredDataRepresentable
-
-extension Bool: StructuredDataRepresentable {
-    public var structuredData: StructuredData {
-        return .bool(self)
-    }
-}
-
-extension Double: StructuredDataRepresentable {
-    public var structuredData: StructuredData {
-        return .double(self)
-    }
-}
-
-extension Int: StructuredDataRepresentable {
-    public var structuredData: StructuredData {
-        return .int(self)
-    }
-}
-
-extension String: StructuredDataRepresentable {
-    public var structuredData: StructuredData {
-        return .string(self)
-    }
-}
-
-extension Data: StructuredDataRepresentable {
-    public var structuredData: StructuredData {
-        return .data(self)
-    }
-}
-
-extension Optional where Wrapped: StructuredDataRepresentable {
-    public var structuredData: StructuredData {
-        switch self {
-        case .some(let wrapped): return wrapped.structuredData
-        case .none: return .null
-        }
-    }
-}
-
-extension Array where Element: StructuredDataRepresentable {
-    public var structuredDataArray: [StructuredData] {
-        return self.map({$0.structuredData})
-    }
-
-    public var structuredData: StructuredData {
-        return .array(structuredDataArray)
-    }
-}
-
-public protocol StructuredDataDictionaryKeyRepresentable {
-    var structuredDataDictionaryKey: String { get }
-}
-extension String: StructuredDataDictionaryKeyRepresentable {
-    public var structuredDataDictionaryKey: String {
-        return self
-    }
-}
-extension Dictionary where Key: StructuredDataDictionaryKeyRepresentable, Value: StructuredDataRepresentable {
-    public var structuredDataDictionary: [String: StructuredData] {
-        var dictionary: [String: StructuredData] = [:]
-
-        for (key, value) in self.map({($0.0.structuredDataDictionaryKey, $0.1.structuredData)}) {
-            dictionary[key] = value
-        }
-
-        return dictionary
-    }
-
-    public var structuredData: StructuredData {
-        return .dictionary(structuredDataDictionary)
-    }
-}
-
-// MARK: StructuredDataFallibleRepresentable
-
-extension Optional : StructuredDataFallibleRepresentable {
-    public func asStructuredData() throws -> StructuredData {
-        if case .some(let wrapped) = self, let representable = wrapped as? StructuredDataFallibleRepresentable {
-            return try representable.asStructuredData()
-        } else if Wrapped.self is StructuredDataFallibleRepresentable {
-            return .null
-        }
-        throw StructuredDataError.notStructuredDataRepresentable(Wrapped.self)
-    }
-}
-
-extension Array : StructuredDataFallibleRepresentable {
-    public func asStructuredData() throws -> StructuredData {
-        var array: [StructuredData] = []
-
-        for element in self {
-            guard let representable = element as? StructuredDataFallibleRepresentable else {
-                throw StructuredDataError.notStructuredDataRepresentable(Element.self)
-            }
-            array.append(try representable.asStructuredData())
-        }
-
-        return .array(array)
-    }
-}
-
-extension Dictionary : StructuredDataFallibleRepresentable {
-    public func asStructuredData() throws -> StructuredData {
-        var dictionary: [String: StructuredData] = [:]
-
-        for (key, value) in self {
-            guard let representable = value as? StructuredDataFallibleRepresentable else {
-                throw StructuredDataError.notStructuredDataRepresentable(Value.self)
-            }
-            guard let key = key as? StructuredDataDictionaryKeyRepresentable else {
-                throw StructuredDataError.notStructuredDataDictionaryKeyRepresentable(Key.self)
-            }
-            dictionary[key.structuredDataDictionaryKey] = try representable.asStructuredData()
-        }
-
-        return .dictionary(dictionary)
-    }
 }
 
 // MARK: Initializers
