+++ Sources/StructuredDataInitializable.swift
@@ -1,8 +1,8 @@
-// Reflection.swift
+// StructuredData.swift
 //
 // The MIT License (MIT)
 //
-// Copyright (c) 2016 Brad Hilton
+// Copyright (c) 2016 Zewo
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
